added new screen successfully
